requirements.txt
Flask==3.0.1
gunicorn==23.0.0

requests==2.32.4
structlog==25.1.0

google-auth==2.3.2
google-cloud-aiplatform
google-cloud-vision
jsonschema
torch>=2.0.0
numpy>=1.24.0
scipy>=1.10.0
